refactor: remove unused dependencies and clean up test code
